+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/model/FrontendScreenAsset.java
@@ -4,15 +4,14 @@ import de.mephisto.vpin.restclient.frontend.FrontendPlayerDisplay;
 import javafx.scene.media.MediaPlayer;
 import javafx.stage.Stage;
 
-import java.io.InputStream;
+import java.net.URL;
 
 public class FrontendScreenAsset {
   private Stage screenStage;
   private int rotation;
   private FrontendPlayerDisplay display;
   private String mimeType;
-  private InputStream inputStream;
-  private String url;
+  private URL url;
   private int duration;
   private String name;
 
@@ -82,19 +81,11 @@ public class FrontendScreenAsset {
     this.mimeType = mimeType;
   }
 
-  public InputStream getInputStream() {
-    return inputStream;
-  }
-
-  public void setInputStream(InputStream inputStream) {
-    this.inputStream = inputStream;
-  }
-
-  public String getUrl() {
+  public URL getUrl() {
     return url;
   }
 
-  public void setUrl(String url) {
+  public void setUrl(URL url) {
     this.url = url;
   }
 }