+++ vpin-studio-commons/src/main/java/de/mephisto/vpin/commons/fx/pausemenu/model/PauseMenuScreensFactory.java
@@ -2,20 +2,17 @@ package de.mephisto.vpin.commons.fx.pausemenu.model;
 
 import de.mephisto.vpin.commons.FrontendScreensManager;
 import de.mephisto.vpin.restclient.client.VPinStudioClient;
-import de.mephisto.vpin.restclient.games.FrontendMediaItemRepresentation;
-import de.mephisto.vpin.restclient.games.FrontendMediaRepresentation;
-import de.mephisto.vpin.restclient.games.GameRepresentation;
 import de.mephisto.vpin.restclient.frontend.FrontendPlayerDisplay;
 import de.mephisto.vpin.restclient.frontend.VPinScreen;
+import de.mephisto.vpin.restclient.games.FrontendMediaItemRepresentation;
+import de.mephisto.vpin.restclient.games.GameRepresentation;
 import edu.umd.cs.findbugs.annotations.NonNull;
-import edu.umd.cs.findbugs.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.InputStream;
 import java.lang.invoke.MethodHandles;
-import java.util.ArrayList;
-import java.util.List;
+import java.net.MalformedURLException;
+import java.net.URL;
 
 /**
  * Factory class to create frontend screens for the assets found in the Popper resource folders.
@@ -55,23 +52,18 @@ public class PauseMenuScreensFactory {
 //    return null;
 //  }
 
-  @Nullable
-  public static FrontendScreenAsset createScreenStage(VPinStudioClient client, GameRepresentation game, FrontendPlayerDisplay display, VPinScreen screen, FrontendMediaItemRepresentation defaultMediaItem) {
-    InputStream imageStream = client.getWheelIcon(game.getId(), false);
-    if (imageStream != null) {
-      FrontendScreenAsset asset = new FrontendScreenAsset();
-      asset.setDisplay(display);
-      asset.setRotation(0);
-      asset.setDuration(0);
-      asset.setMimeType(defaultMediaItem.getMimeType());
-      asset.setInputStream(imageStream);
-      asset.setName(defaultMediaItem.getName());
-      asset.setUrl(client.getURL(defaultMediaItem.getUri()));
+  @NonNull
+  public static FrontendScreenAsset createScreenStage(VPinStudioClient client, GameRepresentation game, FrontendPlayerDisplay display, VPinScreen screen, FrontendMediaItemRepresentation defaultMediaItem, int rotation) throws MalformedURLException {
+    FrontendScreenAsset asset = new FrontendScreenAsset();
+    asset.setDisplay(display);
+    asset.setRotation(rotation);
+    asset.setDuration(0);
+    asset.setMimeType(defaultMediaItem.getMimeType());
+    asset.setName(defaultMediaItem.getName());
+    asset.setUrl(new URL(client.getURL(defaultMediaItem.getUri())));
 
-      FrontendScreensManager.getInstance().showScreen(asset);
-      LOG.info("Created stage for screen " + screen + ", asset \"" + defaultMediaItem.getName() + "\"");
-      return asset;
-    }
-    return null;
+    FrontendScreensManager.getInstance().showScreen(asset);
+    LOG.info("Created stage for screen " + screen + ", asset \"" + defaultMediaItem.getName() + "\"");
+    return asset;
   }
 }